+++ packages/openmrs/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @openfn/language-openmrs
 
+## 4.0.1
+
+### Patch Changes
+
+- Updated dependencies [77a690f]
+  - @openfn/language-common@2.0.2
+
+## 4.0.0
+
+### Major Changes
+
+- Remove `createEncounter` and `createPatient` functions
+- Add pagination support on request helper function
+
+### Minor Changes
+
+- c8dbd21: Add cursor and dateFns helper functions
+
+### Migration Guide
+
+The `createEncounter` and `createPatient` functions have been removed from the
+OpenMRS adaptor. Use the `create` function with the appropriate resource type.
+
+To create a new encounter, you can use the following code:
+
+```js
+create('encounter', $.encounter);
+```
+
+To create a new patient, you can use the following code:
+
+```js
+create('patient', $.patient);
+```
+
 ## 3.1.4
 
 ### Patch Changes